requirements.txt
# 奇点量化交易平台 - Python 依赖
# 使用方式：pip install -r requirements.txt

# 数据处理
numpy>=2.4.2
pandas>=3.0.1
scipy>=1.15.0

# 数据源
akshare>=1.18.35
yfinance>=0.2.0

# AI / 机器学习
openai>=2.26.0
google-cloud-aiplatform>=1.140.0
scikit-learn>=1.6.0

# Google Cloud
google-cloud-bigquery>=3.40.1
google-cloud-language>=2.19.0
google-cloud-storage>=3.9.0
google-cloud-scheduler>=3.15.0

# Azure
azure-cognitiveservices-speech>=1.48.2
azure-cosmos>=4.15.0
azure-identity>=1.19.0

# Web 框架
flask>=3.1.3
gunicorn>=25.1.0

# 爬虫
scrapy>=2.11.0
selenium>=4.28.0
beautifulsoup4>=4.13.0
lxml>=5.3.0

# 可视化
matplotlib>=3.10.8
plotly>=6.0.0

# 工具库
python-dotenv>=1.2.2
requests>=2.31.0
tenacity>=9.0.0
tqdm>=4.67.0

# 邮件
yagmail>=0.15.293

# 监控
datadog>=0.49.0

# 开发工具
pytest>=8.3.0
black>=25.1.0
flake8>=7.1.0
